# Conflicts:
#	src/it/unibo/oop17/ga_game/model/entities/AbstractEnemy.java

+++ src/it/unibo/oop17/ga_game/view/SpriteAnimation.java
@@ -9,15 +9,15 @@ import javafx.util.Duration;
 public class SpriteAnimation extends Transition {
     private final ImageView imageView;
     private final int count;
-    private final int offsetX;
-    private final int offsetY;
-    private final int width;
-    private final int height;
+    private final double offsetX;
+    private final double offsetY;
+    private final double width;
+    private final double height;
 
     private int lastIndex;
 
     public SpriteAnimation(final ImageView imageView, final Duration duration, final int count,
-            final int offsetX, final int offsetY, final int width, final int height) {
+            final double offsetX, final double offsetY, final double width, final double height) {
         this.imageView = imageView;
         this.count = count;
         this.offsetX = offsetX;
@@ -35,8 +35,8 @@ public class SpriteAnimation extends Transition {
 
         final int index = Math.min((int) Math.floor(k * count), count - 1);
         if (index != lastIndex) {
-            final int x = index % cols * width + offsetX;
-            final int y = index % rows * height + offsetY;
+            final double x = index % cols * width + offsetX;
+            final double y = index % rows * height + offsetY;
             imageView.setViewport(new Rectangle2D(x, y, width, height));
             lastIndex = index;
         }
